fix(homepage): validator must run on first page load, not just in lazy api chunk
Bug surfaced during PR-#88 manual smoke test: `vite preview` of a
production bundle with `VITE_API_KEY=hf_dev_key_2026` loaded the home
page cleanly without firing the validator. Reason: api.ts (where the
validator lived) is code-split by Vite into a separate lazy chunk —
DashboardPage / AdminPage / TelemetryRow import it, but the
marketing-style home page does not. A misconfigured production bundle
would therefore render the home page fine and only fast-fail when the
user clicked into the dashboard, defeating the "fast-fail at first
browser load" guarantee the validator was supposed to provide.

Fix:
- Extract DEV_FALLBACK_KEY + validateBuildTimeApiKey + the top-level
  call into a new module `homepage/src/services/api-key-validator.ts`.
  This is the canonical declaration site for both.
- main.tsx adds a side-effect import: `import './services/api-key-validator';`
  Puts the validator's call into the entry chunk (`index-*.js`), which
  loads on every route including `/`. Cost: ~30 bytes in the entry
  chunk + one inlined throw site. Bundle artifact diff confirmed: the
  throw text now lives in `dist/assets/index-*.js`, not the lazy
  `api-*.js` chunk.
- api.ts re-exports DEV_FALLBACK_KEY + validateBuildTimeApiKey so the
  existing unit tests in api-key-validator.test.ts continue to import
  from the same path. The API_KEY local on line 21 still uses
  DEV_FALLBACK_KEY (now via the re-export).
- Cross-file symmetry test (dev-fallback-symmetry.test.ts) now reads
  `homepage/src/services/api-key-validator.ts` rather than api.ts —
  the canonical declaration moved.

Verification:
- Backend tests: 80 pass
- Homepage tests: 45 pass
- `VITE_API_KEY=hf_dev_key_2026 npx vite build` — bundle builds, throw
  text is in `dist/assets/index-*.js` (entry chunk), confirmed via
  `grep "is set (case-insensitively)" dist/assets/*.js`.
- Re-running the manual smoke test (vite preview + open home page in
  browser) should now show the Uncaught Error in DevTools on first
  load, regardless of route.
